moved presentational components to templates
