+++ vis/js/templates/HoverPopover.jsx
