+++ vis/js/templates/SubdisciplineTitle.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 
-import Heading from "./Heading";
-import Backlink from "./Backlink";
-import ContextLine from "./ContextLine";
+import Heading from "../components/Heading";
+import Backlink from "../components/Backlink";
+import ContextLine from "../components/ContextLine";
 
 class SubdisciplineTitle extends React.Component {
   render() {
